Refactoring unit tests
Made the unit tests more modular
